feature(deployment): local context from truffle develop to ganache-cli - temporary instable - problem to dynamically retrieve deployed contracts address. Focus on test net (ropsten) deployment.

+++ demo-mvp/setup.sh
@@ -2,54 +2,55 @@
 
 # clean setup at end of process
 clean () {
-    rm -r ../BTUToken/build
-    rm -r ../RES/build
-    rm src/BTU/BTU.json
-    rm src/RES/RES.json
-    if [ -n "$bc" ] && ps -p $bc > /dev/null; then
-        kill -s 9 $bc
+    if [ -n "$customRPC" ] && ps -p $customRPC > /dev/null; then
+        kill -s 9 $customRPC
     fi
-    if [ -n "$dApp" ] && ps -p $dApp > /dev/null; then
-        kill -s 9 $dApp
-    fi 
 }
 trap clean EXIT
 
 # Check if truffle is installed
 # if not, propose it
-type truffle >/dev/null 2>&1 || { 
-    echo >&2 "Truffle is required but not found.";
-    read -n1 -p "Are you agree to install this package? [y,n]" agreement 
+type truffle >/dev/null 2>&1 || {
+    echo >&2 "Truffle is required but was not found.";
+    read -n1 -p "Do you want install this package? [Y/n]" agreement
     case $agreement in
-        y|Y) npm install -g truffle ;;
-        n|N) "Ok."; exit 1 ;;
-        *) "Ok."; exit 1 ;;
+        ''|y|Y) npm install -g truffle ;;
+        n|N) exit 1 ;;
+        *) exit 1 ;;
     esac
 }
-# compile and deploy BTU contract on local context
-cd ../BTUToken/;
-# install dependencies
-npm install;
-# compile BTU.sol
-truffle compile;
-# launch a local blockchain context using truffle 
-(truffle develop ) &
-bc=$!;
-# migrate BTU contract to local context
-truffle migrate;
-# give to the dApp the good abi
-# repeat process with RES contract
-cd ../RES/;
-npm install;
-truffle compile;
-truffle migrate;
-# setup and install the dApp
-cd ../demo-mvp/;
-mkdir src/BTU/;
-mkdir src/RES/;
-cp ../BTUToken/build/contracts/BTU.json src/BTU/;
+
+# Check if ganache-cli is installed
+# if not, propose it
+type ganache-cli >/dev/null 2>&1 || {
+    echo >&2 "Ganache is required but was not found.";
+    read -n1 -p "Do you want install this package? [Y/n]" agreement
+    case $agreement in
+        ''|y|Y) npm install -g ganache-cli ;;
+        n|N) exit 1 ;;
+        *) exit 1 ;;
+    esac
+}
+
+# launch a local blockchain context using customRPC
+(ganache-cli -p 9545 -m 'candy maple cake sugar pudding cream honey rich smooth crumble sweet treat' >> customRPC.log 2>> customRPC_error.log) & customRPC=$!
+# compile and deploy contracts
+./compile_contract.sh ../BTUToken/ & btuCompilation=$!
+wait $btuCompilation
+btuCompilResult=$?
+./compile_contract.sh ../RES/ & resCompilation=$!
+wait $resCompilation
+resCompilResult=$?
+if [ $btuCompilResult -ne 0 ] && [ $resCompilResult -ne 0 ]; then
+    echo 'Contract compilation failed check logs for errors !'
+    exit 1;
+fi
+
+mkdir src/BTU/
+mkdir src/RES/
+cp ../BTUToken/build/contracts/BTUTokenSale.json src/BTU/;
 cp ../RES/build/contracts/RES.json src/RES/;
+
 npm install;
-npm run start;
-dApp=$!;
-wait $bc
+echo 'DONE'
+wait $customRPC
